requirements.txt
pdfplumber
pandas
python-docx
